@@ -1,6 +1,6 @@
 {
   "name": "cypress",
-  "version": "9.7.0",
+  "version": "12.2.0",
   "main": "index.js",
   "scripts": {
     "postinstall": "node index.js --exec install",
@@ -26,7 +26,7 @@
     "dayjs": "^1.10.4",
     "debug": "^4.3.2",
     "enquirer": "^2.3.6",
-    "eventemitter2": "^6.4.3",
+    "eventemitter2": "6.4.7",
     "execa": "4.1.0",
     "executable": "^4.1.1",
     "extract-zip": "2.0.1",
@@ -54,23 +54,75 @@
     "bin",
     "lib",
     "index.js",
+    "index.mjs",
     "types/**/*.d.ts",
-    "types/net-stubbing.ts"
+    "mount-utils",
+    "vue",
+    "react",
+    "vue2",
+    "react18",
+    "angular",
+    "svelte"
   ],
   "bin": {
     "cypress": "bin/cypress"
   },
   "engines": {
-    "node": ">=12.0.0"
+    "node": "^14.0.0 || ^16.0.0 || >=18.0.0"
   },
   "types": "types",
+  "exports": {
+    ".": {
+      "import": "./index.mjs",
+      "require": "./index.js",
+      "types": "./types/index.d.ts"
+    },
+    "./vue": {
+      "import": "./vue/dist/cypress-vue.esm-bundler.js",
+      "require": "./vue/dist/cypress-vue.cjs.js",
+      "types": "./vue/dist/index.d.ts"
+    },
+    "./vue2": {
+      "import": "./vue2/dist/cypress-vue2.esm-bundler.js",
+      "require": "./vue2/dist/cypress-vue2.cjs.js",
+      "types": "./vue2/dist/index.d.ts"
+    },
+    "./package.json": {
+      "import": "./package.json",
+      "require": "./package.json"
+    },
+    "./react": {
+      "import": "./react/dist/cypress-react.esm-bundler.js",
+      "require": "./react/dist/cypress-react.cjs.js",
+      "types": "./react/dist/index.d.ts"
+    },
+    "./react18": {
+      "import": "./react18/dist/cypress-react.esm-bundler.js",
+      "require": "./react18/dist/cypress-react.cjs.js",
+      "types": "./react18/dist/index.d.ts"
+    },
+    "./mount-utils": {
+      "require": "./mount-utils/dist/index.js",
+      "types": "./mount-utils/dist/index.d.ts"
+    },
+    "./angular": {
+      "import": "./angular/dist/index.js",
+      "require": "./angular/dist/index.js",
+      "types": "./angular/dist/index.d.ts"
+    },
+    "./svelte": {
+      "import": "./svelte/dist/cypress-svelte.esm-bundler.js",
+      "require": "./svelte/dist/cypress-svelte.cjs.js",
+      "types": "./svelte/dist/index.d.ts"
+    }
+  },
   "buildInfo": {
     "commitBranch": "develop",
-    "commitSha": "884e8e963b917eb96922517bbb314446b5495658",
-    "commitDate": "2022-05-23T15:08:23.000Z",
+    "commitSha": "ebf57a82e6d9912e0eac2f0b9f620b950eb5893c",
+    "commitDate": "2022-12-20T01:56:10.000Z",
     "stable": true
   },
-  "description": "Cypress.io end to end testing tool",
+  "description": "Cypress is a next generation front end testing tool built for the modern web",
   "homepage": "https://github.com/cypress-io/cypress",
   "license": "MIT",
   "bugs": {
@@ -88,6 +140,7 @@
     "e2e",
     "end-to-end",
     "integration",
+    "component",
     "mocks",
     "runner",
     "spies",
